+++ src/components/SQForm/SQFormDropdown.js
@@ -110,9 +110,11 @@ function SQFormDropdown({
           );
         })}
       </Select>
-      <FormHelperText error={isFieldError} required={isRequired}>
-        {HelperTextComponent}
-      </FormHelperText>
+      {!isDisabled && (
+        <FormHelperText error={isFieldError} required={isRequired}>
+          {HelperTextComponent}
+        </FormHelperText>
+      )}
     </Grid>
   );
 }